Feat: Migrated Passcode Module to KMP
