+++ libs/mifos-passcode/src/androidMain/AndroidManifest.xml
