--- libs/mifos-passcode/src/main/kotlin/org/mifos/library/passcode/theme/Font.kt
@@ -1,34 +1,0 @@
-/*
- * Copyright 2024 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See https://github.com/openMF/mobile-wallet/blob/master/LICENSE.md
- */
-package org.mifos.library.passcode.theme
-
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontStyle
-import androidx.compose.ui.text.font.FontWeight
-import com.mifos.library.passcode.R
-
-internal val LatoFonts = FontFamily(
-    Font(
-        resId = R.font.lib_mifos_passcode_lato_regular,
-        weight = FontWeight.Normal,
-        style = FontStyle.Normal,
-    ),
-    Font(
-        resId = R.font.lib_mifos_passcode_lato_bold,
-        weight = FontWeight.Bold,
-        style = FontStyle.Normal,
-    ),
-    Font(
-        resId = R.font.lib_mifos_passcode_lato_black,
-        weight = FontWeight.Black,
-        style = FontStyle.Normal,
-    ),
-)
