--- libs/mifos-passcode/src/main/kotlin/org/mifos/library/passcode/utility/VibrationFeedback.kt
@@ -1,48 +1,0 @@
-/*
- * Copyright 2024 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See https://github.com/openMF/mobile-wallet/blob/master/LICENSE.md
- */
-package org.mifos.library.passcode.utility
-
-import android.content.Context
-import android.os.Build
-import android.os.CombinedVibration
-import android.os.VibrationEffect
-import android.os.Vibrator
-import android.os.VibratorManager
-import org.mifos.library.passcode.utility.Constants.VIBRATE_FEEDBACK_DURATION
-
-internal object VibrationFeedback {
-
-    @Suppress("DEPRECATION")
-    internal fun vibrateFeedback(context: Context) {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-            (context.getSystemService(Context.VIBRATOR_MANAGER_SERVICE) as VibratorManager).vibrate(
-                CombinedVibration.createParallel(
-                    VibrationEffect.createOneShot(
-                        VIBRATE_FEEDBACK_DURATION,
-                        VibrationEffect.DEFAULT_AMPLITUDE,
-                    ),
-                ),
-            )
-        } else {
-            (context.getSystemService(Context.VIBRATOR_SERVICE) as Vibrator).let {
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                    it.vibrate(
-                        VibrationEffect.createOneShot(
-                            VIBRATE_FEEDBACK_DURATION,
-                            VibrationEffect.DEFAULT_AMPLITUDE,
-                        ),
-                    )
-                } else {
-                    it.vibrate(VIBRATE_FEEDBACK_DURATION)
-                }
-            }
-        }
-    }
-}
