--- libs/mifos-passcode/src/main/res/drawable/lib_mifos_passcode_ic_delete.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-    Copyright 2024 Mifos Initiative
-
-    This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
-    If a copy of the MPL was not distributed with this file,
-    You can obtain one at https://mozilla.org/MPL/2.0/.
-
-    See https://github.com/openMF/mobile-wallet/blob/master/LICENSE.md
--->
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="18dp"
-    android:viewportWidth="24"
-    android:viewportHeight="18">
-    <path
-        android:fillColor="#000000"
-        android:fillType="evenOdd"
-        android:pathData="M7.2485,0.6685C7.4857,0.4023 7.8253,0.25 8.1818,0.25H20.5909C21.4216,0.25 22.2226,0.5739 22.8167,1.1574C23.4116,1.7416 23.75,2.5389 23.75,3.375V14.625C23.75,15.4611 23.4116,16.2584 22.8167,16.8426C22.2226,17.4261 21.4216,17.75 20.5909,17.75H8.1818C7.8253,17.75 7.4857,17.5977 7.2485,17.3315L0.5667,9.8315C0.1444,9.3576 0.1444,8.6424 0.5667,8.1685L7.2485,0.6685ZM8.7423,2.75L3.1741,9L8.7423,15.25H20.5909C20.7729,15.25 20.943,15.1788 21.065,15.059C21.1861,14.94 21.25,14.7834 21.25,14.625V3.375C21.25,3.2166 21.1861,3.06 21.065,2.941C20.943,2.8212 20.7729,2.75 20.5909,2.75H8.7423ZM11.1082,5.3116C11.5919,4.8191 12.3834,4.812 12.8759,5.2957L14.8636,7.248L16.8514,5.2957C17.3439,4.812 18.1353,4.8191 18.6191,5.3116C19.1028,5.8042 19.0957,6.5956 18.6032,7.0793L16.6475,9L18.6032,10.9207C19.0957,11.4044 19.1028,12.1959 18.6191,12.6884C18.1353,13.1809 17.3439,13.188 16.8514,12.7043L14.8636,10.7521L12.8759,12.7043C12.3834,13.188 11.5919,13.1809 11.1082,12.6884C10.6245,12.1959 10.6316,11.4044 11.1241,10.9207L13.0797,9L11.1241,7.0793C10.6316,6.5956 10.6245,5.8042 11.1082,5.3116Z" />
-</vector>
