+++ mifospay-shared/src/commonMain/kotlin/org/mifospay/shared/navigation/LoginNavGraph.kt
@@ -12,6 +12,7 @@ package org.mifospay.shared.navigation
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.navigation
+import org.mifos.library.passcode.navigateToPasscodeScreen
 import org.mifospay.feature.auth.navigation.LOGIN_ROUTE
 import org.mifospay.feature.auth.navigation.loginScreen
 import org.mifospay.feature.auth.navigation.mobileVerificationScreen
@@ -28,8 +29,7 @@ internal fun NavGraphBuilder.loginNavGraph(navController: NavController) {
     ) {
         loginScreen(
             onNavigateBack = navController::popBackStack,
-            // TODO:: we should navigate to passcode graph
-            onNavigateToPasscodeScreen = navController::navigateToMainGraph,
+            onNavigateToPasscodeScreen = navController::navigateToPasscodeScreen,
             onNavigateToSignupScreen = navController::navigateToSignupMethod,
         )
